requirements.txt
google-api-python-client==2.88.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.100.0
uvicorn==0.22.0
pydantic==2.0.3
pytest==7.3.1
pytest-asyncio==0.21.0
aiohttp==3.8.5
jinja2==3.1.2
itsdangerous==2.1.2 